Add address search by postcode view
A new view has been created for the "address search by postcode" feature, allowing users to find UK addresses using a specific postcode. This view includes both the search form and the display of search results. It also gives users the option to save the preferred addresses for future reference.