Replace full HTML template with layout component
The full HTML template in the show blade for address by postcode has been replaced with a reusable layout component. This replacement is part of a refactor aimed at promoting code reusability and maintaining a consistent website layout across pages. The changes primarily involve removing sections in the template that are now included in the layout component such as html tags, head section, and body tags.