Add check for duplicate addresses
Added a new functionality in the AddressController to prevent duplication of addresses. If a user attempts to enter an address that already exists in the system, an error message will be returned instead of creating a new address entry. The controller now also returns a success status when a new address is successfully created.