Add 'store' method to AddressController
A 'store' method has been added to the AddressController class. This method fetches a new address using an external API and stores it in the database if it doesn't already exist. In case of any errors (like the API request failing, no results, or address already existing in the database), appropriate json responses with status and message are returned.